first commit
This tool will load a Numbas exam from a .zip package, and try answering
every part of every question with the expected answer.
It'll then try to resume that attempt in review mode.

If it succeeds, it prints the JSON dictionary {'success': true}.
If not, it prints a dictionary with 'success': false and a `message`
attribute plus some other stuff.